fix: 🐛 ensure check enough funds
Instead of relying on the mathematical failures for reversion, the first
level ensures that there are sufficient funds in the account (EOA or
stub/service provider), and then as a backup, mathematical failures are
then checked. This further enhances the UX as well with an accurate
failure message instead of an arbitrary mathematical panic.